remove hard coded origin paths
The first version used a hard coded hostnames and paths
for the DASH stream. This version removes that hard coded
values, by wrapping the original addresses into the
path component of the URL.

This commit also adds a new lambda URLCreate that takes a
URL to a DASH manifest and generates a wrapped version that
will point at the ManifestProxy lambda.