use base64 to encode origin URL
When deployed to Azure, the URL is getting partially
un-escaped before it reaches the lambda function, causing
invalid origin URLs to be created when the original URL
escaping is removed.

This commit adds the option to base64 encode the URLs, to
work around this issue. It makes the URLs very long, but
is more robust.