add reverse proxy that can modify HTTP responses from origin
In some cloud environments it is tricky to use lambdas because
they assume that they are being used for an API and therefore
do not support binary responses.

Another issue is that lambdas might be charged by the number of
milliseconds they take to complete. If this lambda has to make
a request to origin, the time to fetch that data would become
part of the chargeable time.

This commit adds an HTTP reverse proxy (in Python) that has the
same logic as the lambdas and can make origin requests and modify
its responses.

It can be used as a custom origin behind a CDN, so that the CDN
provides caching the clients and the proxy acts as a "man in the
middle" between the CDN and the real origin.